requirements.txt
python-dotenv
discord.py
requests
isodate
